Split serializers into separate files
